Refactor pagination utils into separate files
