+++ routes/api/posts/create.js
@@ -1,7 +1,7 @@
 const authenticate = require('../../../utils/authenticate');
 const redisClient = require('../../../config/redisClient');
 const validatePostInput = require('../../../validation/post');
-const { easyParse } = require('../../../utils/pagination');
+const easyParse = require('../../../utils/easyParse');
 const { v4: uuidv4 } = require('uuid');
 const Post = require('../../../models/Post');
