+++ utils/easyParse.js
@@ -1,0 +1,5 @@
+const easyParse = (item) => {
+	return typeof item === 'string' ? JSON.parse(item) : item;
+};
+
+module.exports = easyParse;
