+++ utils/padWithZeros.js
@@ -1,0 +1,5 @@
+const padWithZeros = (number, length) => {
+	return number.toString().padStart(length, '0');
+};
+
+module.exports = padWithZeros;
